Revamp to be like makefile.mingw.in, make the MSVC build actually work
2000-10-06  Tor Lillqvist  <tml@iki.fi>

	* makefile.msc.in: Revamp to be like makefile.mingw.in, make
	the MSVC build actually work again.

	* gmodule/makefile.msc.in
	* gobject/makefile.msc.in
	* gthread/makefile.msc.in: New files, like their mingw counterparts.

	* gmodule/Makefile.am
	* gobject/Makefile.am
	* gthread/Makefile.am: Make and distribute them.

	* */makefile.mingw.in: Allow override of GLib version number from
	the build/win32/module.defs file.

	* glib.def: Add new entry point.

	Fixes from Hans Breuer:

	* glib.h (struct DIR): Keep the last readdir result cached inside
	the DIR struct, to enable several DIRs being open simultaneously.

	* gwin32.c (g_win32_readdir): Use the above instead of static.

	* giowin32.c (g_io_channel_win32_make_pollfd): Insert cast to keep
	MSVC happy.

+++ glib/giowin32.c
@@ -897,7 +897,7 @@ g_io_channel_win32_make_pollfd (GIOChannel   *channel,
   if (win32_channel->data_avail_event == NULL)
     create_events (win32_channel);
   
-  fd->fd = win32_channel->data_avail_event;
+  fd->fd = (gint) win32_channel->data_avail_event;
   fd->events = condition;
 
   if (win32_channel->thread_id == 0)
